review(acp): extract KIMCHI_PROVIDER_ID to a leaf module; dedupe test fixtures
Break the flow.ts↔models.ts cycle instead of hardcoding the provider id:
KIMCHI_PROVIDER_ID (and isKimchiProvider) now live in dependency-free
src/kimchi-provider.ts; flow.ts, login/index.ts, startup-auth.ts,
login-command-patch.ts, models.ts, server.ts, and auth-status.ts import
directly from it, and models.ts's two "kimchi-dev" literals are gone.

Also fold the two review flagged test duplications:

- the on-disk models.json seed in server.test.ts becomes one
  seedModelsJson() helper used by both authenticate describes;
- the copy-pasted FakeAgentSession/modelRegistry scaffold in the
  steering and set_session_title suites moves to
  src/modes/acp/__mocks__/fake-agent-session.ts, per the shared-mocks
  guideline.
